pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agridecisionenv-v3"
version = "3.0.0"
description = "Sustainable Farming Decision Environment for reinforcement learning"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "openai>=1.0",
    "openenv-core",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
server = "server.app:main"

[project.entry-points."openenv"]
server = "server.app:app"

[project.optional-dependencies]
dev = ["pytest>=7.0"]
